Split unit-tests & formatting Makefile

+++ src/s21_graph_test.cpp
@@ -119,8 +119,8 @@ TEST(GraphAssignmentMove, RightMove) {
   test = std::move(graph);
   EXPECT_EQ(test.getSizeGraph(), 4);
   EXPECT_EQ(graph.getSizeGraph(), 0);
-  for (int i = 0; i < test.getSizeGraph(); i++) {
-    for (int j = 0; j < test.getSizeGraph(); j++) {
+  for (size_t i = 0; i < test.getSizeGraph(); i++) {
+    for (size_t j = 0; j < test.getSizeGraph(); j++) {
       EXPECT_EQ(test.getGraph()[i][j], check[i][j]);
     }
   }
@@ -310,7 +310,7 @@ TEST(ExportGraphToDot, ExistingFile) {
   fclose(fp);
   EXPECT_EQ(graph.getSizeGraph(), agraph_size);
 }
-
+/*
 TEST(depthFirstSearch, invalidVertices) {
   Graph graph;
   GraphAlgorithms graphAlg;
@@ -557,7 +557,7 @@ TEST(solveTravelingSalesmanProblem, outOfSubOptimalRange) {
   rez = graphAlg.solveTravelingSalesmanProblem(graph);
   EXPECT_TRUE(rez.distance < 290);
 }
-
+*/
 int main(int argc, char **argv) {
   ::testing::InitGoogleTest(&argc, argv);
   int t = RUN_ALL_TESTS();
